DL/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn
python-multipart
ultralytics
pytesseract
opencv-python-headless
Pillow
